Fix clippy warnings in tests
